+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 0.0.3
+* updated FSharp.Data.Adaptive
+* improved CList/CMap diffing
+
 ### 0.0.2
 * generator rewrite